ref(js): Consistent import usage in routes.tsx
Previously these were imported at the top of the file. We can be more
consistent and instead use a dynamic import, but have webpack bundle
that together with whatever chunk the routes are part of.

We can clearly denote what's going on here with a comment, which is an
improvement over the how it is now